+++ src/test/java/org/mule/modules/sqs/SQSConnectorTest.java
@@ -1,11 +1,7 @@
 /**
- * Mule Amazon SQS Connector
- *
- * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
- *
- * The software in this package is published under the terms of the CPAL v1.0
- * license, a copy of which has been included with this distribution in the
- * LICENSE.txt file.
+ * (c) 2003-2015 MuleSoft, Inc. The software in this package is
+ * published under the terms of the CPAL v1.0 license, a copy of which
+ * has been included with this distribution in the LICENSE.md file.
  */
 
 /**
@@ -26,7 +22,7 @@ public class SQSConnectorTest extends FunctionalTestCase {
     }
 
     // TODO: Unit Test each Connector method.
-    
+
     @Ignore("Rewrite this test! It was trying to hit an actual Amazon instance.")
     @Test
     public void testSendMessage() throws Exception {